requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
click==7.1.2
colorama==0.4.4
Flask==1.1.2
gunicorn==20.1.0
importlib-metadata==3.10.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-decouple==3.4
python-dotenv==0.17.0
python-environ==0.4.54
regex==2021.4.4
scikit-learn==0.24.1
scipy==1.6.2
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
